requirements.txt
streamlit==1.32.0
opencv-python==4.9.0.80
numpy==1.26.4
tensorflow==2.15.0
pillow==10.2.0
flask==3.0.2 